requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
transformers
torch
pandas
scikit-learn
beautifulsoup4
requests
spacy
wikipedia-api
